[Login][Game] теперь есть возможность работать с клиентом версии 0.5.101406 от 23.04.2013 в котором есть dedicated сервер и редактор уровней - не все пакеты проверены

+++ AAEmu.Game/Core/Packets/G2C/SCLootBagDataPacket.cs
@@ -24,7 +24,7 @@ namespace AAEmu.Game.Core.Packets.G2C
             {
                 stream.Write(item.TemplateId);
                 stream.Write(item.Id);
-                stream.Write(item.Grade);
+                stream.Write((byte)item.Grade);
                 stream.Write((byte)0);
                 stream.Write(item.Count);
                 stream.Write(item.DetailType);
